Rename run.py to register.py
